mailer/requirements.txt
psycopg2-binary==2.9.5
sshtunnel==0.4.0
Jinja2==3.1.2
